pyproject.toml
[tool.poetry]
name = "fast api"
version = "0.1.0"
description = ""
authors = ["Thiaog"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.114.0"}
ruff = "^0.5.1"
pydantic = {extras = ["email"], version = "^2.9.0"}
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
sqlalchemy = "^2.0.34"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
httpx = "^0.27.0"

[tool.ruff]
line-length = 79
extend-exclude =['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[toll.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . & ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev projeto_fast_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=projeto_fast_api -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
